refactor: optimized the product card
Deleted wishlist card and added the wishlist functionality to product card itself.
This helps in avoiding using another card with almost same code but just change in click handlers. Helps follow DRY principle.
