+++ src/components/index.js
@@ -5,3 +5,4 @@ export { Filters } from "./Filters";
 export { Products } from "./Products";
 export { CartSummary } from "./CartSummary";
 export { CartItems } from "./CartItems";
+export { ProductCard } from "./ProductCard";
